feat(routing): change the defult landing page
